increment wakes at the end of the active state
if the top of the minute comes during the active mode we will capture
the duration correctly but the wake count will be zero, leading to the
possibility that the next minute has a duration but no wake count. Post
increment prevents this case, so if continuous motion is present there
will be one wakeup per minute and 60 seconds of duration.